Split lounge-specific implementation from DrawableRoom
