+++ osu.Game/Screens/OnlinePlay/Lounge/Components/DrawableRoom.cs
@@ -1,32 +1,19 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
-using System.Collections.Generic;
-using osu.Framework;
 using osu.Framework.Allocation;
-using osu.Framework.Audio;
-using osu.Framework.Audio.Sample;
 using osu.Framework.Bindables;
-using osu.Framework.Extensions;
 using osu.Framework.Extensions.Color4Extensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Colour;
 using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Cursor;
 using osu.Framework.Graphics.Effects;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
-using osu.Framework.Graphics.UserInterface;
-using osu.Framework.Input.Bindings;
-using osu.Framework.Input.Events;
 using osu.Game.Beatmaps;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Containers;
 using osu.Game.Graphics.Sprites;
-using osu.Game.Graphics.UserInterface;
-using osu.Game.Graphics.UserInterfaceV2;
-using osu.Game.Input.Bindings;
 using osu.Game.Online.Rooms;
 using osu.Game.Overlays;
 using osu.Game.Screens.OnlinePlay.Components;
@@ -35,104 +22,17 @@ using osuTK.Graphics;
 
 namespace osu.Game.Screens.OnlinePlay.Lounge.Components
 {
-    public class DrawableRoom : OsuClickableContainer, IStateful<SelectionState>, IFilterable, IHasContextMenu, IHasPopover, IKeyBindingHandler<GlobalAction>
+    public class DrawableRoom : CompositeDrawable
     {
-        public const float SELECTION_BORDER_WIDTH = 4;
-        private const float corner_radius = 10;
-        private const float transition_duration = 60;
+        protected const float CORNER_RADIUS = 10;
         private const float height = 100;
 
-        public event Action<SelectionState> StateChanged;
-
-        protected override HoverSounds CreateHoverSounds(HoverSampleSet sampleSet) => new HoverSounds();
-
-        private Drawable selectionBox;
-
-        [Resolved(canBeNull: true)]
-        private LoungeSubScreen loungeScreen { get; set; }
+        public readonly Room Room;
 
         [Resolved]
         private BeatmapManager beatmaps { get; set; }
 
-        [Resolved(canBeNull: true)]
-        private Bindable<Room> selectedRoom { get; set; }
-
-        [Resolved(canBeNull: true)]
-        private LoungeSubScreen lounge { get; set; }
-
-        public readonly Room Room;
-
-        private SelectionState state;
-
-        private Sample sampleSelect;
-        private Sample sampleJoin;
-
-        public SelectionState State
-        {
-            get => state;
-            set
-            {
-                if (value == state)
-                    return;
-
-                state = value;
-
-                if (selectionBox != null)
-                {
-                    if (state == SelectionState.Selected)
-                        selectionBox.FadeIn(transition_duration);
-                    else
-                        selectionBox.FadeOut(transition_duration);
-                }
-
-                StateChanged?.Invoke(State);
-            }
-        }
-
-        public IEnumerable<string> FilterTerms => new[] { Room.Name.Value };
-
-        private bool matchingFilter;
-
-        public bool MatchingFilter
-        {
-            get => matchingFilter;
-            set
-            {
-                matchingFilter = value;
-
-                if (!IsLoaded)
-                    return;
-
-                if (matchingFilter)
-                    this.FadeIn(200);
-                else
-                    Hide();
-            }
-        }
-
-        private int numberOfAvatars = 7;
-
-        public int NumberOfAvatars
-        {
-            get => numberOfAvatars;
-            set
-            {
-                numberOfAvatars = value;
-
-                if (recentParticipantsList != null)
-                    recentParticipantsList.NumberOfCircles = value;
-            }
-        }
-
-        public bool FilteringActive { get; set; }
-
-        protected readonly Container ButtonsContainer = new Container
-        {
-            Anchor = Anchor.CentreRight,
-            Origin = Anchor.CentreRight,
-            RelativeSizeAxes = Axes.Y,
-            AutoSizeAxes = Axes.X
-        };
+        protected Container ButtonsContainer { get; private set; }
 
         private readonly Bindable<RoomCategory> roomCategory = new Bindable<RoomCategory>();
         private readonly Bindable<bool> hasPassword = new Bindable<bool>();
@@ -149,7 +49,7 @@ namespace osu.Game.Screens.OnlinePlay.Lounge.Components
             Height = height;
 
             Masking = true;
-            CornerRadius = corner_radius;
+            CornerRadius = CORNER_RADIUS;
             EdgeEffect = new EdgeEffectParameters
             {
                 Type = EdgeEffectType.Shadow,
@@ -159,9 +59,9 @@ namespace osu.Game.Screens.OnlinePlay.Lounge.Components
         }
 
         [BackgroundDependencyLoader]
-        private void load(OverlayColourProvider colours, AudioManager audio)
+        private void load(OverlayColourProvider colours)
         {
-            Children = new Drawable[]
+            InternalChildren = new Drawable[]
             {
                 // This resolves internal 1px gaps due to applying the (parenting) corner radius and masking across multiple filling background sprites.
                 new Box
@@ -183,7 +83,7 @@ namespace osu.Game.Screens.OnlinePlay.Lounge.Components
                     {
                         RelativeSizeAxes = Axes.Both,
                         Masking = true,
-                        CornerRadius = corner_radius,
+                        CornerRadius = CORNER_RADIUS,
                         Children = new Drawable[]
                         {
                             new GridContainer
@@ -303,7 +203,13 @@ namespace osu.Game.Screens.OnlinePlay.Lounge.Components
                                 },
                                 Children = new Drawable[]
                                 {
-                                    ButtonsContainer,
+                                    ButtonsContainer = new Container
+                                    {
+                                        Anchor = Anchor.CentreRight,
+                                        Origin = Anchor.CentreRight,
+                                        RelativeSizeAxes = Axes.Y,
+                                        AutoSizeAxes = Axes.X
+                                    },
                                     recentParticipantsList = new RecentParticipantsList
                                     {
                                         Anchor = Anchor.CentreRight,
@@ -316,47 +222,12 @@ namespace osu.Game.Screens.OnlinePlay.Lounge.Components
                         },
                     },
                 },
-                new StatusColouredContainer(transition_duration)
-                {
-                    RelativeSizeAxes = Axes.Both,
-                    Child = selectionBox = new Container
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                        Alpha = state == SelectionState.Selected ? 1 : 0,
-                        Masking = true,
-                        CornerRadius = corner_radius,
-                        BorderThickness = SELECTION_BORDER_WIDTH,
-                        BorderColour = Color4.White,
-                        Child = new Box
-                        {
-                            RelativeSizeAxes = Axes.Both,
-                            Alpha = 0,
-                            AlwaysPresent = true
-                        }
-                    }
-                },
-            };
-
-            sampleSelect = audio.Samples.Get($@"UI/{HoverSampleSet.Default.GetDescription()}-select");
-            sampleJoin = audio.Samples.Get($@"UI/{HoverSampleSet.Submit.GetDescription()}-select");
-        }
-
-        protected override IReadOnlyDependencyContainer CreateChildDependencies(IReadOnlyDependencyContainer parent)
-        {
-            return new CachedModelDependencyContainer<Room>(base.CreateChildDependencies(parent))
-            {
-                Model = { Value = Room }
             };
         }
 
         protected override void LoadComplete()
         {
             base.LoadComplete();
-
-            if (matchingFilter)
-                this.FadeInFromZero(transition_duration);
-            else
-                Alpha = 0;
 
             roomCategory.BindTo(Room.Category);
             roomCategory.BindValueChanged(c =>
@@ -371,57 +242,26 @@ namespace osu.Game.Screens.OnlinePlay.Lounge.Components
             hasPassword.BindValueChanged(v => passwordIcon.Alpha = v.NewValue ? 1 : 0, true);
         }
 
-        public Popover GetPopover() => new PasswordEntryPopover(Room) { JoinRequested = lounge.Join };
-
-        public MenuItem[] ContextMenuItems => new MenuItem[]
+        protected override IReadOnlyDependencyContainer CreateChildDependencies(IReadOnlyDependencyContainer parent)
         {
-            new OsuMenuItem("Create copy", MenuItemType.Standard, () =>
+            return new CachedModelDependencyContainer<Room>(base.CreateChildDependencies(parent))
             {
-                lounge?.Open(Room.DeepClone());
-            })
-        };
-
-        public bool OnPressed(GlobalAction action)
-        {
-            if (selectedRoom.Value != Room)
-                return false;
-
-            switch (action)
-            {
-                case GlobalAction.Select:
-                    TriggerClick();
-                    return true;
-            }
-
-            return false;
+                Model = { Value = Room }
+            };
         }
 
-        public void OnReleased(GlobalAction action)
+        private int numberOfAvatars = 7;
+
+        public int NumberOfAvatars
         {
-        }
-
-        protected override bool ShouldBeConsideredForInput(Drawable child) => state == SelectionState.Selected || child is HoverSounds;
-
-        protected override bool OnClick(ClickEvent e)
-        {
-            if (Room != selectedRoom.Value)
+            get => numberOfAvatars;
+            set
             {
-                sampleSelect?.Play();
-                selectedRoom.Value = Room;
-                return true;
+                numberOfAvatars = value;
+
+                if (recentParticipantsList != null)
+                    recentParticipantsList.NumberOfCircles = value;
             }
-
-            if (Room.HasPassword.Value)
-            {
-                sampleJoin?.Play();
-                this.ShowPopover();
-                return true;
-            }
-
-            sampleJoin?.Play();
-            lounge?.Join(Room, null);
-
-            return base.OnClick(e);
         }
 
         private class RoomNameText : OsuSpriteText
@@ -506,53 +346,6 @@ namespace osu.Game.Screens.OnlinePlay.Lounge.Components
                         Size = new Vector2(14),
                     }
                 };
-            }
-        }
-
-        public class PasswordEntryPopover : OsuPopover
-        {
-            private readonly Room room;
-
-            public Action<Room, string> JoinRequested;
-
-            public PasswordEntryPopover(Room room)
-            {
-                this.room = room;
-            }
-
-            private OsuPasswordTextBox passwordTextbox;
-
-            [BackgroundDependencyLoader]
-            private void load()
-            {
-                Child = new FillFlowContainer
-                {
-                    Margin = new MarginPadding(10),
-                    Spacing = new Vector2(5),
-                    AutoSizeAxes = Axes.Both,
-                    Direction = FillDirection.Horizontal,
-                    Children = new Drawable[]
-                    {
-                        passwordTextbox = new OsuPasswordTextBox
-                        {
-                            Width = 200,
-                        },
-                        new TriangleButton
-                        {
-                            Width = 80,
-                            Text = "Join Room",
-                            Action = () => JoinRequested?.Invoke(room, passwordTextbox.Text)
-                        }
-                    }
-                };
-            }
-
-            protected override void LoadComplete()
-            {
-                base.LoadComplete();
-
-                Schedule(() => GetContainingInputManager().ChangeFocus(passwordTextbox));
-                passwordTextbox.OnCommit += (_, __) => JoinRequested?.Invoke(room, passwordTextbox.Text);
             }
         }
     }
